my account addresses work

+++ django_ecommerce/users/views.py
@@ -4,8 +4,8 @@ from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django_ecommerce.settings import MEDIA_ROOT
-from .forms import UserRegisterForm, UserUpdateForm, ProfileUpdateForm, ProfileUpdateAddressForm
-from .models import Profile
+from .forms import UserRegisterForm, UserUpdateForm, ProfileUpdateForm, UserUpdateAddressForm
+from .models import Profile, UserAddress
 
 
 def register(request):
@@ -27,25 +27,35 @@ def profile(request):
     if request.method == "POST":
         u_form = UserUpdateForm(request.POST, instance=request.user)
         p_form = ProfileUpdateForm(request.POST, request.FILES, instance=request.user.profile)
-        p_address_form = ProfileUpdateAddressForm(request.POST, instance=request.user.profile)
-        if u_form.is_valid() and p_form.is_valid():
+        u_address_form = UserUpdateAddressForm(request.POST, instance=request.user)
+        if u_form.is_valid() and p_form.is_valid() and u_address_form.is_valid():
             u_form.save()
             if general_current_user_img != '/media/images/user/default.jpg' and general_current_user_img != request.user.profile.image.url:
                 os.unlink(os.path.join(MEDIA_ROOT, 'images', 'user', 'profile_pics', general_current_user_img.split('/')[-1]))
             p_form.save()
-            p_address_form.save()
+            u_address_form = UserAddress()
+            u_address_form.user = request.user
+            u_address_form.company = request.POST.get('company')
+            u_address_form.phone_number = request.POST.get('phone_number')
+            u_address_form.address_1 = request.POST.get('address_1')
+            u_address_form.address_2 = request.POST.get('address_2')
+            u_address_form.city = request.POST.get('city')
+            u_address_form.country = request.POST.get('country')
+            u_address_form.postal_code = request.POST.get('postal_code')
+            u_address_form.fax = request.POST.get('fax')
+            u_address_form.save()
             messages.success(request, f'Your account has been updated!')
             return redirect('my-account')
 
     else:
         u_form = UserUpdateForm(instance=request.user)
         p_form = ProfileUpdateForm(instance=request.user.profile)
-        p_address_form = ProfileUpdateAddressForm(instance=request.user.profile)
+        u_address_form = UserUpdateAddressForm()
 
     context = {
         'u_form': u_form,
         'p_form': p_form,
-        'p_address_form': p_address_form,
+        'u_address_form': u_address_form,
         'title': 'My Account',
     }
